Move trigger services tests from st2reactor to st2common
